pyproject.toml
[tool.poetry]
name = "nonebot_plugin_cchess"
version = "0.4.0"
description = "Nonebot2 象棋插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-cchess"
repository = "https://github.com/noneplugin/nonebot-plugin-cchess"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.2.0"
nonebot-plugin-alconna = ">=0.38.2,<1.0.0"
nonebot-plugin-session = "^0.3.0"
nonebot-plugin-userinfo = "^0.2.0"
nonebot-plugin-orm = ">=0.7.0,<1.0.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
nonebot-plugin-orm = { version = ">=0.7.0,<1.0.0", extras = ["default"] }

[tool.pyright]
pythonVersion = "3.8"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
